ChatHistory -> TextHistory
It would be nice to reuse this for things like editing memory, AN, etc.

If we treat *every* serializable change as an undoable action (see the
Command pattern), we can make the save files log-structured, allowing
for continuous autosave, preservation of undo history across save/load,
and effectively "free" serialization.